move GltfPipeline to Source/Scene and add it to .eslintignore
