+++ Source/Scene/GltfJsonLoader.js
@@ -5,12 +5,12 @@ import getJsonFromTypedArray from "../Core/getJsonFromTypedArray.js";
 import getMagic from "../Core/getMagic.js";
 import isDataUri from "../Core/isDataUri.js";
 import Resource from "../Core/Resource.js";
-import addDefaults from "../ThirdParty/GltfPipeline/addDefaults.js";
-import addPipelineExtras from "../ThirdParty/GltfPipeline/addPipelineExtras.js";
-import ForEach from "../ThirdParty/GltfPipeline/ForEach.js";
-import parseGlb from "../ThirdParty/GltfPipeline/parseGlb.js";
-import removePipelineExtras from "../ThirdParty/GltfPipeline/removePipelineExtras.js";
-import updateVersion from "../ThirdParty/GltfPipeline/updateVersion.js";
+import addDefaults from "./GltfPipeline/addDefaults.js";
+import addPipelineExtras from "./GltfPipeline/addPipelineExtras.js";
+import ForEach from "./GltfPipeline/ForEach.js";
+import parseGlb from "./GltfPipeline/parseGlb.js";
+import removePipelineExtras from "./GltfPipeline/removePipelineExtras.js";
+import updateVersion from "./GltfPipeline/updateVersion.js";
 import when from "../ThirdPartyNpm/when.js";
 import ResourceLoader from "./ResourceLoader.js";
 import ResourceLoaderState from "./ResourceLoaderState.js";
