+++ Source/Scene/GltfPipeline/addDefaults.js
