+++ Source/Scene/GltfPipeline/addExtensionsRequired.js
