+++ Source/Scene/GltfPipeline/addPipelineExtras.js
