+++ Source/Scene/GltfPipeline/findAccessorMinMax.js
