+++ Source/Scene/GltfPipeline/forEachTextureInMaterial.js
