+++ Source/Scene/GltfPipeline/getAccessorByteStride.js
