+++ Source/Scene/GltfPipeline/getComponentReader.js
