+++ Source/Scene/GltfPipeline/moveTechniquesToExtension.js
