+++ Source/Scene/GltfPipeline/numberOfComponentsForType.js
