+++ Source/Scene/GltfPipeline/removeExtensionsRequired.js
