+++ Source/Scene/GltfPipeline/removeExtensionsUsed.js
