+++ Source/Scene/GltfPipeline/removePipelineExtras.js
