+++ Source/Scene/GltfPipeline/removeUnusedElements.js
