+++ Source/Scene/GltfPipeline/updateAccessorComponentTypes.js
