+++ Source/Scene/GltfPipeline/updateVersion.js
