+++ Source/Scene/Model.js
@@ -43,14 +43,14 @@ import Texture from "../Renderer/Texture.js";
 import TextureMinificationFilter from "../Renderer/TextureMinificationFilter.js";
 import TextureWrap from "../Renderer/TextureWrap.js";
 import VertexArray from "../Renderer/VertexArray.js";
-import addDefaults from "../ThirdParty/GltfPipeline/addDefaults.js";
-import addPipelineExtras from "../ThirdParty/GltfPipeline/addPipelineExtras.js";
-import ForEach from "../ThirdParty/GltfPipeline/ForEach.js";
-import getAccessorByteStride from "../ThirdParty/GltfPipeline/getAccessorByteStride.js";
-import usesExtension from "../ThirdParty/GltfPipeline/usesExtension.js";
-import numberOfComponentsForType from "../ThirdParty/GltfPipeline/numberOfComponentsForType.js";
-import parseGlb from "../ThirdParty/GltfPipeline/parseGlb.js";
-import updateVersion from "../ThirdParty/GltfPipeline/updateVersion.js";
+import addDefaults from "./GltfPipeline/addDefaults.js";
+import addPipelineExtras from "./GltfPipeline/addPipelineExtras.js";
+import ForEach from "./GltfPipeline/ForEach.js";
+import getAccessorByteStride from "./GltfPipeline/getAccessorByteStride.js";
+import usesExtension from "./GltfPipeline/usesExtension.js";
+import numberOfComponentsForType from "./GltfPipeline/numberOfComponentsForType.js";
+import parseGlb from "./GltfPipeline/parseGlb.js";
+import updateVersion from "./GltfPipeline/updateVersion.js";
 import when from "../ThirdPartyNpm/when.js";
 import Axis from "./Axis.js";
 import BlendingState from "./BlendingState.js";
